Speed limit feature

+++ src/main/java/br/com/viniciusfinger/geofences/model/Fence.java
@@ -32,7 +32,7 @@ public class Fence {
     @JoinColumn(name = "fence_id")
     private List<FenceFeatureInterest> features;
 
-    private Integer speedLimit;
+    private Float speedLimit;
 
     @Pattern(regexp = "^#([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$", message = "Invalid color hex code")
     private String colorHex;
